+++ lab1/task1/__pycache__/func.cpython-37.pyc
