+++ lab1/task1/caesar.py
@@ -1,18 +1,16 @@
-from func import *
+import argparse
+from utils.func import *
 
 
 
 def encrypt_caesar(text: str, shift: int, alphabet: str) -> str:
     """
-    Шифрует текст алгоритмом Цезаря
+    Шифрует текст алгоритмом Цезаря с заданным сдвигом.
 
-    Args:
-        text: Текст для шифрования
-        shift: Величина сдвига (ключ)
-        alphabet: Используемый алфавит
-
-    Returns:
-        Зашифрованный текст
+    :param encrypted_text: Текст для дешифрования
+    :param shift: Величина сдвига
+    :param alphabet: Используемый алфавит
+    :return: Зашифрованная строка
     """
     encrypted = []
     for char in text.upper():
@@ -24,7 +22,13 @@ def encrypt_caesar(text: str, shift: int, alphabet: str) -> str:
     return ''.join(encrypted)
 
 def decrypt_caesar(encrypted_text: str, shift: int, alphabet: str) -> str:
-    """Дешифрует текст алгоритмом Цезаря с заданным сдвигом"""
+    """
+    Дешифрует текст алгоритмом Цезаря с заданным сдвигом
+    :param encrypted_text: Текст для дешифрования
+    :param shift: Величина сдвига
+    :param alphabet: Используемый алфавит
+    :return: Зашифрованная строка
+    """
     decrypted = []
     for char in encrypted_text.upper():
         if char in alphabet:
@@ -34,29 +38,47 @@ def decrypt_caesar(encrypted_text: str, shift: int, alphabet: str) -> str:
             decrypted.append(char)
     return ''.join(decrypted)
 
+def parse_args():
+    """
+    Парсинг аргументов командной строки.
+    :return:
+    """
+    parser = argparse.ArgumentParser(
+        description="Программа для шифрования/дешифрования методом Цезаря"
+    )
+    parser.add_argument(
+        'mode',
+        choices=['encrypt', 'decrypt'],
+        help='Режим работы: encrypt - шифрование, decrypt - дешифрование'
+    )
+    parser.add_argument(
+        '-c', '--config',
+        default='task1/config.json',
+        help='Путь к файлу конфигурации (по умолчанию: task1/config.json)'
+    )
+    return parser.parse_args()
+
 
 
 def main():
+    args = parse_args()
     try:
-        config = open_json("task1/config.json")
-        text = open_file(config["original_text"])
-        key = open_json(config["key"])
-
-        encrypted_text = encrypt_caesar(text, key["shift"], config["alphabet"])
-        save_file(config["encrypted_text"], encrypted_text)
+        config = open_json(args.config)
+        if args.mode == 'encrypt':
+            text = open_file(config["original_text"])
+            key = open_json(config["key"])
+            result = encrypt_caesar(text, key["shift"], config["alphabet"])
+            save_file(config["encrypted_text"], result)
+            print("Текст успешно зашифрован")
+        elif args.mode == 'decrypt':
+            text = open_file(config["encrypted_text"])
+            key = open_json(config["key"])
+            result = decrypt_caesar(text, key["shift"], config["alphabet"])
+            save_file(config["decrypted_text"], result)
+            print("Текст успешно расшифрован")
+        return 0
     except Exception as e:
-        print(f"Ошибка : {e} ")
-    try:
-        config = open_json("task1/config.json")
-        text = open_file(config["encrypted_text"])
-        key = open_json(config["key"])
-
-        decrypted_text = decrypt_caesar(encrypted_text, key["shift"], config["alphabet"])
-        save_file(config["decrypted_text"], decrypted_text)
-    except Exception as e:
-        print(f"Ошибка : {e} ")
-    return 0
-
+        print(f"Ошибка: {e}")
 
 if __name__ == "__main__":
     main()
