+++ lab1/task2/__pycache__/func.cpython-37.pyc
