--- task1/decrypted_text.txt
@@ -1,2 +1,0 @@
-ХОРОШИХ КНИГ О ШКОЛЕ НАПИСАНО НЕМАЛО. НО МНЕ ИНТЕРЕСНО БЫЛО БЫ ИССЛЕДОВАТЬ ОДНУ ЛИНИЮ В ЭТОЙ ОГРОМНОЙ ТЕМЕ: ЧТО ОСТАЕТСЯ ОТ УЧИТЕЛЯ, ОТ КЛАССА В ХАРАКТЕРЕ ЧЕЛОВЕКА, КРОМЕ ЗНАНИЙ. В СУЩНОСТИ, У КАЖДОГО ВЗРОСЛОГО ЕСТЬ СВОИ ВПЕЧАТЛЕНИЯ, СВОИ ВОСПОМИНАНИЯ О ШКОЛЕ. ПЛОХИЕ ИЛИ ХОРОШИЕ. ОДНИ УЧИТЕЛЯ ПОМНЯТСЯ, ДРУГИЕ НЕТ. ОДНИ КЛАССНЫЕ ТОВАРИЩИ ВРЕЗАЛИСЬ В ПАМЯТЬ, ДРУГИЕ ЗАБЫЛИСЬ. И НЕ ВСЕГДА ЭТО МОЖНО ОБЪЯСНИТЬ СТЕПЕНЬЮ ДРУЖБЫ. НЕТ, ТУТ ДЕЙСТВУЮТ ИНЫЕ, ГЛУБОКО СКРЫТЫЕ, ОЧЕВИДНО, СЛОЖНЫЕ ПРИЧИНЫ.
-КРОМЕ ТАЛАНТЛИВЫХ УЧИТЕЛЕЙ ЕСТЬ И ДРУГИЕ, ЛЮБИМЫЕ УЧИТЕЛЯ. НАВЕРНОЕ, ЛЮБИМЫХ УЧИТЕЛЕЙ БОЛЬШЕ, ЧЕМ ТАЛАНТЛИВЫХ. НАВЕРНОЕ, МОЖНО ЗАСЛУЖИТЬ ЛЮБОВЬ, ДАЖЕ НЕ ИМЕЯ ОТПУЩЕННЫХ ПРИРОДОЙ ПЕДАГОГИЧЕСКИХ СПОСОБНОСТЕЙ. ТАЛАНТЛИВЫМ СТАТЬ НЕЛЬЗЯ, А ВОТ ЛЮБИМЫМ СТАТЬ МОЖНО.
