--- task1/encrypted_text.txt
@@ -1,2 +1,0 @@
-ФНПНЧЗФЯЙМЗВЯНЯЧЙНКДЯМ ОЗР МНЯМДЛ КН.ЯМНЯЛМДЯЗМСДПДРМНЯАЪКНЯАЪЯЗРРКДГНБ СЫЯНГМТЯКЗМЗЭЯБЯЬСНИЯНВПНЛМНИЯСДЛД:ЯЦСНЯНРС ДСРЮЯНСЯТЦЗСДКЮ,ЯНСЯЙК РР ЯБЯФ П ЙСДПДЯЦДКНБДЙ ,ЯЙПНЛДЯЖМ МЗИ.ЯБЯРТШМНРСЗ,ЯТЯЙ ЕГНВНЯБЖПНРКНВНЯДРСЫЯРБНЗЯБОДЦ СКДМЗЮ,ЯРБНЗЯБНРОНЛЗМ МЗЮЯНЯЧЙНКД.ЯОКНФЗДЯЗКЗЯФНПНЧЗД.ЯНГМЗЯТЦЗСДКЮЯОНЛМЮСРЮ,ЯГПТВЗДЯМДС.ЯНГМЗЯЙК РРМЪДЯСНБ ПЗШЗЯБПДЖ КЗРЫЯБЯО ЛЮСЫ,ЯГПТВЗДЯЖ АЪКЗРЫ.ЯЗЯМДЯБРДВГ ЯЬСНЯЛНЕМНЯНАЩЮРМЗСЫЯРСДОДМЫЭЯГПТЕАЪ.ЯМДС,ЯСТСЯГДИРСБТЭСЯЗМЪД,ЯВКТАНЙНЯРЙПЪСЪД,ЯНЦДБЗГМН,ЯРКНЕМЪДЯОПЗЦЗМЪ.
-ЙПНЛДЯС К МСКЗБЪФЯТЦЗСДКДИЯДРСЫЯЗЯГПТВЗД,ЯКЭАЗЛЪДЯТЦЗСДКЮ.ЯМ БДПМНД,ЯКЭАЗЛЪФЯТЦЗСДКДИЯАНКЫЧД,ЯЦДЛЯС К МСКЗБЪФ.ЯМ БДПМНД,ЯЛНЕМНЯЖ РКТЕЗСЫЯКЭАНБЫ,ЯГ ЕДЯМДЯЗЛДЮЯНСОТШДММЪФЯОПЗПНГНИЯОДГ ВНВЗЦДРЙЗФЯРОНРНАМНРСДИ.ЯС К МСКЗБЪЛЯРС СЫЯМДКЫЖЮ,Я ЯБНСЯКЭАЗЛЪЛЯРС СЫЯЛНЕМН.
