--- task1/key.json
@@ -1,3 +1,0 @@
-{
-  "shift": 164
-}
