--- task1/original_text.txt
@@ -1,2 +1,0 @@
-Хороших книг о школе написано немало. Но мне интересно было бы исследовать одну линию в этой огромной теме: что остается от учителя, от класса в характере человека, кроме знаний. В сущности, у каждого взрослого есть свои впечатления, свои воспоминания о школе. Плохие или хорошие. Одни учителя помнятся, другие нет. Одни классные товарищи врезались в память, другие забылись. И не всегда это можно объяснить степенью дружбы. Нет, тут действуют иные, глубоко скрытые, очевидно, сложные причины.
-Кроме талантливых учителей есть и другие, любимые учителя. Наверное, любимых учителей больше, чем талантливых. Наверное, можно заслужить любовь, даже не имея отпущенных природой педагогических способностей. Талантливым стать нельзя, а вот любимым стать можно.
