--- task2/cod13.txt
@@ -1,10 +1,0 @@
-$dF-BY!tM=8-VBU-VOOdABME>8d-3B9hABh98-IY>>8$-AUBU98dU!d3QdLMEYCB-!83B98W-IU3BhQ-A-VtdKd>BYK
-U>M-U3>UEY>8->Y-M3QUtGJUEY>MM-$dFOh>A=MM-AUBU9YPQ9dU!9YJhdB-AtCLM-E-M>IdA38-KY33MEY
-U3>UE>8K-Q9dMKhnd3BEUK-$dF-BY!tM=-PEtPdB3P-M$-3QU3U!>U3BGU!d3QdLMEYBG-!83B98W-QUM3A-IU!YEtd>Md-M-hIYtd>Md-VtdKd>BUE
-AtY338-$dF-BY!tM=-KURhB-9YJtMLYBG3P-QU-3QU3U!h-U!9Y!UBAMAUttMJMW-3MBhY=MW-AURIY-IEY-AtCLY-$dFM9hCB3P-E-UIM>-MBUB-Zd-M>IdA3
-3hnd3BEhCB-IEY-U3>UE>8$-KdBUIY-=dQULdL>YP-YI9d3Y=MP-MUBA98BYP-YI9d3Y=MP
-Q9M-=dQULdL>UW-YI9d3Y=MM-AYZI8W-VtdKd>B-KY33MEY-$9Y>MB338tAh->Y-3QM3UA-VtdKd>BUE-AUBU98d-MKdCB-UIM>YAUE8W-$dF
-VBU-QUJEUtPdB-IU!YEtPBG->UE8d-VtdKd>B8-!dJ->dU!$UIMKU3BMQd9d9Y3Q9dIdtPBG-3hnd3BEhCnMd
-UBA98BYP-YI9d3Y=MP-E-3EUC-ULd9dIG-B9d!hdB-QUM3AY-3tdIhCndRU3EU!UI>URU-M>IdA3Y-E-3thLYd-AUttMJMM-LBU-KUZdB-Q9MEd3BM-A3>MZd>MC-Q9UMJEUIMBdtG>U3BM-Q9M-E83UAUW-JYR9hJAdBY!tM=8
-E8!U9-KdBUIY-JYEM3MB-UB-B9d!UEY>MW-A-!83B9UIdW3BEMC-MU!xdKh-IY>>8$
-A9UKd-BURU-$dFBY!tM=8-KURhB-!8BG-IM>YKMLd3AMKM-LBUQUJEUtPdB-MJKd>PBG-M$-9YJKd9-E-Q9U=d33d-9Y!UB8-U!d3QdLMEYPBdK-3YK8K-UQBMKMJY=MC-QYKPBM-M-Q9UMJEUIMBdtG>U3BM
