--- task2/config2.json
@@ -1,7 +1,0 @@
-{
-    "freq_ru": "freq_ru.json",
-    "real_frequency": "real_frequency.json",
-    "cod13_txt": "cod13.txt",
-    "encrypted_text": "encrypted.txt",
-    "key2": "key2.json"
-}
