--- task2/key2.json
@@ -1,35 +1,0 @@
-{
-    "-": " ",
-    "U": "О",
-    "D": "И",
-    "M": "Е",
-    "B": "А",
-    "Y": "Н",
-    "3": "Т",
-    ">": "С",
-    "E": "Р",
-    "9": "В",
-    "T": "М",
-    "8": "Л",
-    "A": "Д",
-    "I": "Я",
-    "K": "К",
-    "!": "П",
-    "Q": "З",
-    "H": "Ы",
-    "P": "Ь",
-    "J": "У",
-    "=": "Ч",
-    "L": "Ж",
-    "$": "Г",
-    "C": "Х",
-    "G": "Ф",
-    "W": "Й",
-    "\n": "Ю",
-    "V": "Б",
-    "F": "Ц",
-    "R": "Ш",
-    "N": "Щ",
-    "Z": "Э",
-    "O": "Ъ"
-}
