--- task2/real_frequency.json
@@ -1,36 +1,0 @@
-{
-    "$": 0.012974976830398516,
-    "d": 0.07970342910101946,
-    "F": 0.006487488415199258,
-    "-": 0.11121408711770157,
-    "B": 0.06765523632993513,
-    "Y": 0.057460611677479144,
-    "!": 0.025023169601482854,
-    "t": 0.028730305838739572,
-    "M": 0.07506950880444857,
-    "=": 0.013901760889712697,
-    "8": 0.028730305838739572,
-    "V": 0.0074142724745134385,
-    "U": 0.08433734939759036,
-    "O": 0.0027803521779425394,
-    "A": 0.028730305838739572,
-    "E": 0.03614457831325301,
-    ">": 0.03892493049119555,
-    "3": 0.057460611677479144,
-    "9": 0.03614457831325301,
-    "h": 0.020389249304911955,
-    "I": 0.027803521779425393,
-    "Q": 0.025023169601482854,
-    "L": 0.013901760889712697,
-    "C": 0.011121408711770158,
-    "W": 0.009267840593141797,
-    "K": 0.026876737720111215,
-    "\n": 0.008341056533827619,
-    "G": 0.010194624652455977,
-    "J": 0.014828544949026877,
-    "P": 0.017608897126969416,
-    "n": 0.004633920296570899,
-    "R": 0.006487488415199258,
-    "Z": 0.0037071362372567192,
-    "x": 0.0009267840593141798
-}
